WIP working on supporting multiple files uploaded/downloaded

+++ server/file_controller.go
@@ -31,23 +31,20 @@ func (ctlr *FileController) Upload(res http.ResponseWriter, req *http.Request) s
 	key := crypto.GenerateKey(32)
 
 	// encrypt the file
-	fileId, err := crypto.EncryptMultipartReader(mr, req.ContentLength, key)
+	fileIdGrpoup, err := crypto.EncryptMultipartReader(mr, req.ContentLength, key)
 
-	//fmt.Println("fileName: ", fileName)
-	//fmt.Println("contentType:", contentType)
-
-	// marshal the fileId to a []byte of json data
-	marshaledFileId, err := json.Marshal(fileId)
+	// marshal the fileIdGroup to a []byte of json data
+	marshaledFileIdGroup, err := json.Marshal(fileIdGroup)
 	if err != nil {
 		fmt.Println(err)
 		return err.Error()
 	}
 
 	// Convert []byte to string
-	serializedFileId := base64.URLEncoding.EncodeToString(marshaledFileId)
+	serializedFileIdGroup := base64.URLEncoding.EncodeToString(marshaledFileIdGroup)
 
 	// build the url to be returned making sure it gets encoded
-	urlToMAC := ctlr.buildDownloadUrl(serializedFileId)
+	urlToMAC := ctlr.buildDownloadUrl(serializedFileIdGroup)
 	mac := crypto.ComputeMAC(urlToMAC, ctlr.masterKey)
 
 	//fmt.Printf("URL MAC: %q\n", mac)
@@ -66,61 +63,65 @@ func (ctlr *FileController) Upload(res http.ResponseWriter, req *http.Request) s
 }
 
 func (ctlr *FileController) Download(params martini.Params, res http.ResponseWriter, req *http.Request) {
-	serializedFileId := params["fileId"]
+	serializedFileIdGroup := params["fileId"]
 	mac := params["mac"]
 
-	//fmt.Println("serializedFileId: ", serializedFileId)
+	//fmt.Println("serializedFileIdGroup: ", serializedFileIdGroup)
 	//fmt.Println("mac: ", mac)
 
-	message := ctlr.buildDownloadUrl(serializedFileId)
+	message := ctlr.buildDownloadUrl(serializedFileIdGroup)
 	if !crypto.CheckMAC(message, mac, ctlr.masterKey) {
 		panic("FAILURE! URL may have been corrupted or tampered with.")
 	}
 
-	jsonToUnmarshal, err := base64.URLEncoding.DecodeString(serializedFileId)
+	jsonToUnmarshal, err := base64.URLEncoding.DecodeString(serializedFileIdGroup)
 	if err != nil {
 		panic("server error")
 	}
 
-	var fileId model.FileIdentifier
-	err = json.Unmarshal(jsonToUnmarshal, &fileId)
+	var fileIdGroup model.FileIdentifierGroup
+	err = json.Unmarshal(jsonToUnmarshal, &fileIdGroup)
 	if err != nil {
 		panic("server error")
 	}
 
-	//fmt.Println("fileId: ", fileId)
-	file := fileId.Path
+	//fmt.Println("fileIdGroup: ", fileIdGroup)
 
-	// decode the key back to a []byte
-	//key, err := base64.URLEncoding.DecodeString(fileId.Key)
-	//if err != nil {
-	//	panic("server error")
-	//}
+	if len(fileIdGroup.FileIds) == 1 {
+		file := fileIdGroup.Path
 
-	//set the relevant headers.
-	res.Header().Set("Content-Disposition", "attachment; filename="+fileId.FileName)
-	res.Header().Set("Content-Type", fileId.ContentType)
+		//set the relevant headers.
+		res.Header().Set("Content-Disposition", "attachment; filename="+fileId.FileName)
+		res.Header().Set("Content-Type", fileId.ContentType)
 
-	err = crypto.DecryptStream(res, file, fileId.Key, fileId.Checksum)
-	if err != nil {
-		panic(err.Error())
-	}
+		err = crypto.DecryptStream(res, file, fileId.Key, fileId.Checksum)
+		if err != nil {
+			panic(err.Error())
+		}
 
-	err = os.Remove(file)
+		err = os.Remove(file)
 
-	if err != nil {
-		//fmt.Println(err)
-		//panic(err)
-		return
+		if err != nil {
+			//fmt.Println(err)
+			//panic(err)
+			return
+		}
+	} else if len(fileIdGroup.FileIds) > 1 {
+		// TODO:  Allow user to download all files as a zip if there is more than one.
+		//	      otherwise, send the single file to the user with the correct ContentType.
+		fmt.Println("blah")
+	} else {
+		//TODO: return error
+		fmt.Println("blah")
 	}
 
 	return
 }
 
-func (ctlr *FileController) buildDownloadUrl(serializedFileId string) string {
+func (ctlr *FileController) buildDownloadUrl(serializedFileIdGroup string) string {
 	// build the url to be returned making sure it gets encoded
 	var baseUrl *url.URL
-	baseUrl, err := baseUrl.Parse("http://privfiles.com:3000/download/" + serializedFileId)
+	baseUrl, err := baseUrl.Parse("http://privfiles.com:3000/download/" + serializedFileIdGroup)
 	if err != nil {
 		panic("server error")
 	}
